requirements.txt
Django~=3.0.6
djangorestframework~=3.11.0
markdown~=3.2.2
django-filter~=2.2.0